requirements.txt
Flask==3.1.0
Flask-SQLAlchemy==3.1.1
PyMySQL==1.1.1
cryptography==43.0.3   # required by PyMySQL for auth plugins
python-dotenv==1.0.1   # optional: load .env for DATABASE_URL
